refac: move handler test to folder
