+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Tests/Application/MeteringPoints/Handlers/MeteringPointCreatedHandlerTests.cs
@@ -16,7 +16,6 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using AutoFixture.Xunit2;
-using GreenEnergyHub.Charges.Application;
 using GreenEnergyHub.Charges.Application.MeteringPoints.Handlers;
 using GreenEnergyHub.Charges.Domain.MeteringPointCreatedEvents;
 using GreenEnergyHub.Charges.Domain.MeteringPoints;
@@ -28,7 +27,7 @@ using NodaTime;
 using Xunit;
 using Xunit.Categories;
 
-namespace GreenEnergyHub.Charges.Tests.Application.MeteringPoints
+namespace GreenEnergyHub.Charges.Tests.Application.MeteringPoints.Handlers
 {
     [UnitTest]
     public class MeteringPointCreatedHandlerTests
